neural_network/agent_based_modeling.py
import numpy as np
from tools.quantum_circuit_designer import QuantumCircuitDesigner

class AgentBasedModeling:
    def __init__(self, agents, environment):
        self.agents = agents
        self.environment = environment
        self.qc_designer = QuantumCircuitDesigner()
    
    def __del__(self):
        if hasattr(self, 'qc_designer'):
            self.qc_designer.cleanup()  # Assuming cleanup method exists

    def simulate(self, steps):
        for _ in range(steps):
            for agent in self.agents:
                agent.act(self.environment, self.agents)
            self.environment.update(self.agents)

class QuantumAgent:
    def __init__(self, state, strategy):
        self.state = state
        self.strategy = strategy

    def act(self, environment, agents):
        self.state = self.strategy(self.state, environment, agents)

    def communicate(self, other_agent):
        # Example communication method
        pass

    def cooperate(self, other_agent):
        # Example cooperation method
        pass

    def compete(self, other_agent):
        # Example competition method
        pass

class QuantumEnvironment:
    def __init__(self, initial_conditions):
        self.state = initial_conditions

    def update(self, agents):
        # Update the environment state based on some rules and agent interactions
        pass

class NeutrosophicLogic:
    def __init__(self, truth, indeterminacy, falsity):
        self.truth = truth
        self.indeterminacy = indeterminacy
        self.falsity = falsity

    def apply(self, data):
        # Apply neutrosophic logic to the data
        return self.truth * data + self.indeterminacy * (1 - data) - self.falsity * data

class QuantumDataFiltration:
    def __init__(self, neutrosophic_logic):
        self.neutrosophic_logic = neutrosophic_logic

    def filter(self, data):
        return self.neutrosophic_logic.apply(data)
