neural_network/tests/test_agent_based_modeling.py
import unittest
import numpy as np
from neural_network.agent_based_modeling import AgentBasedModeling, QuantumAgent, QuantumEnvironment

class TestAgentBasedModeling(unittest.TestCase):

    def setUp(self):
        self.agents = [
            QuantumAgent(state=np.random.rand(4), strategy=self.dummy_strategy),
            QuantumAgent(state=np.random.rand(4), strategy=self.dummy_strategy)
        ]
        self.environment = QuantumEnvironment(initial_conditions=np.random.rand(4))
        self.agent_based_modeling = AgentBasedModeling(agents=self.agents, environment=self.environment)

    def dummy_strategy(self, state, environment, agents):
        return state + np.random.rand(4)

    def test_simulate(self):
        initial_states = [agent.state.copy() for agent in self.agents]
        self.agent_based_modeling.simulate(steps=10)
        for initial_state, agent in zip(initial_states, self.agents):
            self.assertFalse(np.array_equal(initial_state, agent.state), "Agent state should change after simulation")

    def test_agent_interactions(self):
        agent1, agent2 = self.agents
        agent1.communicate(agent2)
        agent1.cooperate(agent2)
        agent1.compete(agent2)
        self.assertTrue(True)  # Add appropriate assertions based on the expected behavior

    def test_environment_update(self):
        initial_state = self.environment.state.copy()
        self.environment.update(self.agents)
        self.assertFalse(np.array_equal(initial_state, self.environment.state), "Environment state should change after update")

if __name__ == '__main__':
    unittest.main()
